AMBARI-25006. Move AddServiceRequest to addservice, Service/Component/Host to top-level
